+++ distant-core/src/manager/data.rs
@@ -1,3 +1,6 @@
+mod capabilities;
+pub use capabilities::*;
+
 mod destination;
 pub use destination::*;