+++ distant-core/src/manager/data/request.rs
@@ -1,13 +1,38 @@
 use super::{ChannelId, ConnectionId, Destination};
 use crate::{DistantMsg, DistantRequestData, Map};
+use derive_more::IsVariant;
 use distant_net::Request;
 use serde::{Deserialize, Serialize};
+use strum::{EnumDiscriminants, EnumIter, EnumMessage, EnumString};
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, EnumDiscriminants, Serialize, Deserialize)]
 #[cfg_attr(feature = "clap", derive(clap::Subcommand))]
+#[strum_discriminants(derive(
+    strum::Display,
+    EnumIter,
+    EnumMessage,
+    EnumString,
+    Hash,
+    PartialOrd,
+    Ord,
+    IsVariant,
+    Serialize,
+    Deserialize
+))]
+#[cfg_attr(
+    feature = "schemars",
+    strum_discriminants(derive(schemars::JsonSchema))
+)]
+#[strum_discriminants(name(ManagerCapabilityKind))]
+#[strum_discriminants(strum(serialize_all = "snake_case"))]
 #[serde(rename_all = "snake_case", deny_unknown_fields, tag = "type")]
 pub enum ManagerRequest {
+    /// Retrieve information about the server's capabilities
+    #[strum_discriminants(strum(message = "Supports retrieving capabilities"))]
+    Capabilities,
+
     /// Launch a server using the manager
+    #[strum_discriminants(strum(message = "Supports launching distant on remote servers"))]
     Launch {
         // NOTE: Boxed per clippy's large_enum_variant warning
         destination: Box<Destination>,
@@ -18,6 +43,7 @@ pub enum ManagerRequest {
     },
 
     /// Initiate a connection through the manager
+    #[strum_discriminants(strum(message = "Supports connecting to remote servers"))]
     Connect {
         // NOTE: Boxed per clippy's large_enum_variant warning
         destination: Box<Destination>,
@@ -29,6 +55,7 @@ pub enum ManagerRequest {
 
     /// Opens a channel for communication with a server
     #[cfg_attr(feature = "clap", clap(skip))]
+    #[strum_discriminants(strum(message = "Supports opening a channel with a remote server"))]
     OpenChannel {
         /// Id of the connection
         id: ConnectionId,
@@ -36,6 +63,9 @@ pub enum ManagerRequest {
 
     /// Sends data through channel
     #[cfg_attr(feature = "clap", clap(skip))]
+    #[strum_discriminants(strum(
+        message = "Supports sending data through a channel with a remote server"
+    ))]
     Channel {
         /// Id of the channel
         id: ChannelId,
@@ -47,21 +77,26 @@ pub enum ManagerRequest {
 
     /// Closes an open channel
     #[cfg_attr(feature = "clap", clap(skip))]
+    #[strum_discriminants(strum(message = "Supports closing a channel with a remote server"))]
     CloseChannel {
         /// Id of the channel to close
         id: ChannelId,
     },
 
     /// Retrieve information about a specific connection
+    #[strum_discriminants(strum(message = "Supports retrieving connection-specific information"))]
     Info { id: ConnectionId },
 
     /// Kill a specific connection
+    #[strum_discriminants(strum(message = "Supports killing a remote connection"))]
     Kill { id: ConnectionId },
 
     /// Retrieve list of connections being managed
+    #[strum_discriminants(strum(message = "Supports retrieving a list of managed connections"))]
     List,
 
     /// Signals the manager to shutdown
+    #[strum_discriminants(strum(message = "Supports being shut down on demand"))]
     Shutdown,
 }