+++ tests/cli/mod.rs
@@ -1,5 +1,6 @@
 mod action;
 mod fixtures;
+mod manager;
 mod repl;
 mod scripts;
 mod utils;